+++ Changelog.md
@@ -14,13 +14,15 @@
 
 - [#609]: Added `Writer::write_serializable` to provide the capability to serialize
   arbitrary types using serde when using the lower-level `Writer` API.
+- [#615]: Added ability to set entity resolver when deserialize using borrowing reader.
 
 ### Bug Fixes
 
 ### Misc Changes
 
 
-[#609]: https://github.com/tafia/quick-xml/issues/609
+[#609]: https://github.com/tafia/quick-xml/pull/609
+[#615]: https://github.com/tafia/quick-xml/pull/615
 
 
 ## 0.29.0 -- 2023-06-13